feat: move replay to ba subcommand
Closes https://github.com/ByteArena/bytearena/issues/244
