+++ ba/action/replay/main.go
@@ -1,7 +1,6 @@
-package main
+package replay
 
 import (
-	"flag"
 	"fmt"
 	"strconv"
 	"time"
@@ -9,31 +8,26 @@ import (
 	notify "github.com/bitly/go-notify"
 	"github.com/skratchdot/open-golang/open"
 
+	mapcmd "github.com/bytearena/bytearena/ba/action/map"
 	"github.com/bytearena/bytearena/common"
+	"github.com/bytearena/bytearena/common/mappack"
 	"github.com/bytearena/bytearena/common/recording"
 	"github.com/bytearena/bytearena/common/utils"
 	"github.com/bytearena/bytearena/vizserver"
 	"github.com/bytearena/bytearena/vizserver/types"
 )
 
-func main() {
-	filename := flag.String("file", "", "Filename")
-	port := flag.Int("viz-port", 8080, "Specifiy the port of the visualization server")
-
-	flag.Parse()
-
-	utils.Assert(*filename != "", "--file must be set")
-
+func Main(filename string, port int) {
 	game := NewMockGame(10)
 
-	vizserver := NewVizService(*port, game, *filename)
+	vizserver := NewVizService(port, game, filename)
 	// Below line is used to serve assets locally
 	// TODO(jerome): find a way to bundle the trainer with the assets
 	//vizserver.SetPathToAssets("/Users/jerome/Code/other/assets/")
 
 	vizserver.Start()
 
-	url := "http://localhost:" + strconv.Itoa(*port) + "/record/1"
+	url := "http://localhost:" + strconv.Itoa(port) + "/record/1"
 
 	fmt.Println("\033[0;34m\nReplay ready; open " + url + " in your browser.\033[0m\n")
 	open.Run(url)
@@ -62,9 +56,19 @@ func NewVizService(port int, game *MockGame, recordFile string) *vizserver.VizSe
 	vizgames := make([]*types.VizGame, 1)
 	vizgames[0] = types.NewVizGame(game)
 
-	vizservice := vizserver.NewVizService("0.0.0.0:"+strconv.Itoa(port), webclientpath, func() ([]*types.VizGame, error) {
-		return vizgames, nil
-	}, recordStore)
+	mappack, errMappack := mappack.UnzipAndGetHandles(mapcmd.GetMapLocation(mapName))
+	if errMappack != nil {
+		utils.FailWith(errMappack)
+	}
+
+	vizservice := vizserver.NewVizService(
+		"0.0.0.0:"+strconv.Itoa(vizport),
+		webclientpath,
+		mapName,
+		func() ([]*types.VizGame, error) { return vizgames, nil },
+		recorder,
+		mappack,
+	)
 
 	return vizservice
 }
